[iOS] Use local prefs for SetUpList with Customization
crrev.com/c/5757240 changed set_up_list/utils to use the user prefs
instead of the local prefs since the customization prefs should be
specific to the user. However, the other checks within the set-up list
should continue to use the local prefs since they're tied to the device.
This way, each profile can choose to toggle the visibility of the
set-up list, but the content within it will be generalized to the
device.

Bug: 350990359
Change-Id: I6913c15d68c91d0b1e5f5fc619c05520a2d866e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5761693
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Adam Arcaro <adamta@google.com>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337421}