Fix precision loss on preferred refresh rates
DisplayPrivate::SetSupportedRefreshRates is called with a vector of
floats, which are subsequently converted to TimeDelta. Eventually the
chosen preferred interval is converted back to float and sent via
DisplayPrivate::SetPreferredRefreshRate. Over the course of those two
conversions, there is a chance to lose some precision such that the
eventual preference does not belong to the list of supported rates.

This CL adds a map to RootCompositorFrameSinkImpl which stores the
precise rates when they are received, keyed by their corresponding
TimeDelta conversions. This allows the precise values to be looked up
for the final preference result, preventing the possibility of precision
loss and eliminating the need to convert a second time.

Bug: b:328599219
Change-Id: I820d3b8a1092da7b5a3da75b74ba3fd52ff48c36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5738951
Commit-Queue: Andrew Wolfers <aswolfers@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337432}