Replace variable-length arrays with FixedArrays
Variable-length arrays are not part of the C++ standard, and are
dangerous to use because they allocate a potentially-unbounded
amount of data on the stack. In addition, they trigger a clang warning
(-Wvla-extension) which is currently suppressed in Chromium builds.

This change replaces VLAs with base::FixedArray, which has similar
performance but does not risk overflowing the stack (arrays which are
too large will be allocated on the heap instead).

Bug: 349656479
Change-Id: I86435370113e0d89647746151883534035cb0319
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5752549
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1337448}