Remove uses of variable-length arrays
Variable-length arrays are not part of the C++ standard, and are
dangerous to use because they allocate a potentially-unbounded
amount of data on the stack. In addition, they trigger a clang warning
(-Wvla-extension) which is currently suppressed in Chromium builds.

This simply removes a warning by marking the length of the array as
`constexpr`, allowing the compiler to figure out that its length is
static.

Bug: 349656479
Change-Id: I81c814a22015906728d97256394673129de9bbd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5748293
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Siddhartha S <ssid@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337458}