IDB: Fix flakiness in LevelDBScopeTest.BrokenDBForCleanup
The test intends to break the test DB after commit but before the
cleanup task starts running. But the cleanup task runs on a separate
thread and could complete execution before we break the DB, and hence
the flakiness.

This CL fixes this by adding a callback to `LevelDBScopes::Commit()`
that runs after commit but before the cleanup task is scheduled. The
test now uses this callback to reliably break the DB.

The CL also makes some drive-by fixes in the other tests to ensure we
validate expectations of teardown status correctly.

Validated that the flakiness is fixed by running the tests with high
parallelism locally.

Bug: 350130537
Change-Id: I159cc1eecc2102ba9fb83a653da99dbb347e442d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5758999
Auto-Submit: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Ayu Ishii <ayui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337510}