Update HTTPS-First Mode settings to include Balanced Mode
This replaces the previously prototyped HFM-in-Incognito tri-state
settings UI with a new toggle+selection UI. A main toggle button
controls whether to show a group of radio buttons that allow the user
to choose between "Strict Mode" (the existing HFM) and the new
"Balanced Mode" (where we try to show fewer warnings). When the main
toggle is off, the radio buttons are hidden and both strict and
balanced mode prefs are disabled. The new settings UI includes new
strings (although they are not finalized yet).

This also updates the generated_https_first_mode_pref.cc logic to map
between the three settings states (ENABLED_FULL, ENABLED_BALANCED, and
DISABLED) and the two prefs (kHttpsOnlyModeEnabled and
kHttpsFirstBalancedMode).

This also hooks up the kHttpsFirstBalancedMode pref to control whether
balanced mode interstitials be triggered, now that it is
user-controllable.

Finally, this replaces the previous HFM-in-Incognito pref tests (which
are obsolete) to instead test that the Balanced Mode pref/setting is
correctly logged.

All of this is gated behind the kHttpsFirstBalancedMode feature flag.

Screenshot:
https://drive.google.com/file/d/1jl8rAcSRLTEd-gVkRilooPIgiBWRmMmE/view?usp=sharing

Bug: 349860796
Change-Id: Ic848c7fc1f28b2df492bb375f993f975cd5b3fea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5758682
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337527}