[Extensions][UI] Change extensions webstore UI links.
Before this change:

1) the chrome://extensions/?id=<ext_id> page had a "View in Chrome Web
Store" webstore link out that sent the user to the extension's detail
page on the webstore
2) if an extension didn't specify "homepage_url" in its manifest the
URL defaulted to the webstore detail page for the extension's ID. This
was accessed by right clicking the extension badge in the toolbar and
clicking the "<extension_name>" (first item in the context list).

After this change both these UI surfaces will now direct the user to the
chromewebstore.google.com domain instead of the chrome.google.com domain
for the webstore URLs.

Bug: 40073814
Change-Id: I7687022b1d8b91cf9f8e2b85c5cad649c2d01b2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5762919
Reviewed-by: Emily Shack <emshack@chromium.org>
Commit-Queue: Justin Lulejian <jlulejian@chromium.org>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337536}