[Extensions] Change worker registration test observer method call logic.
Before this change
ServiceWorkerTaskQueue::TestObserver::OnWorkerRegistered() was only
called  very early in
ServiceWorkerTaskQueue::DidRegisterServiceWorker(). This makes it
difficult to test the multiple histograms in this method since some are
emitted after OnWorkerRegistered() is called.

After this change we now always call OnWorkerRegistered() during
DidRegisterServiceWorker() so that it can be relied upon for waiting for
histograms to be emitted.

Bug: 346732739
Change-Id: I83ff3e162c6fb6027bcff7786e6dce4c8d52f584
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5761809
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Commit-Queue: Justin Lulejian <jlulejian@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337580}