Remove ReadAnythingTabHelper.
This CL is a refactor with no intended behavior change.

The class ReadAnythingTabHelper is a wrapper around
ReadAnythingSidePanelController that adds nothing. This CL deletes
ReadAnythingTabHelper and replaces all callsites with
ReadAnythingSidePanelController. This CL moves instantiation of the
class to TabFeatures.

Bug: 355485153
Change-Id: I04fce7a5713cdaa70f58314d044e04ba15cc1ea7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5745734
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1337624}