[views-ax] Migrates kDefaultActionVerb attribute to new system in OptionButton
This CL migrates the kDefaultActionVerb value in OptionButtonBase
view whenever its value should change, rather than querying the value
and computing it only when needed.

This CL eliminates the direct assignment of the kDefaultActionVerb
attribute within the View::GetAccessibleNodeData method. Instead, it
introduces a mechanism to update the kDefaultActionVerb attribute in the
accessibility cache when any attribute that contributes to the
accessible kDefaultActionVerb undergoes a change. This approach ensures
that the accessible kDefaultActionVerb remains up-to-date in the
accessibility cache.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: I3ad503dc64ee6dccbfa33cec444f5a60fed6ad35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5705649
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Ashish Kumar <ashishkum@microsoft.com>
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1337675}