[views-ax] Migrates kDefaultActionVerb attribute to new system in CheckButton
This CL migrates the kDefaultActionVerb value in
GlanceablesTaskView::CheckButton view whenever its value should change,
rather than querying the value and computing it only when needed.

This CL eliminates the direct assignment of the kDefaultActionVerb
attribute within the View::GetAccessibleNodeData method. Instead, it
introduces a mechanism to update the kDefaultActionVerb attribute in the
accessibility cache when any attribute that contributes to the
accessible kDefaultActionVerb undergoes a change. This approach ensures
that the accessible kDefaultActionVerb remains up-to-date in the
accessibility cache.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: I8a4ed9b8b429357637717bdc1a33bc200f9a2949
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5705768
Reviewed-by: Javier Contreras <javiercon@microsoft.com>
Commit-Queue: Ashish Kumar <ashishkum@microsoft.com>
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337682}