Limit the input to the language detection model.
The model only considers the first 128 characters of the input. We pass the full text of the page to the language detection model, expecting it to ignore the characters it does not use but the running time is proportional to the full length of the string. 100k of English takes 13ms on my cloudtop.

This could be causing a little jank, especially on slower devices. See
the bug for UMA stats. It's not likely to be a big issue.

The limit is enabled by default for callers of `Predict` but it is controlled by a flag for the translate calls for safe rollout and to
allow for measuring the impact.

This also adds a UMA histogram to time the per-pageload usage of the
TFLite model.


Bug: 352636753
Change-Id: Ia6c3fa4a1ed823681ce3b451b747722ef04a71c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5691065
Commit-Queue: Fergal Daly <fergal@chromium.org>
Auto-Submit: Fergal Daly <fergal@chromium.org>
Commit-Queue: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Trevor Perrier <perrier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337723}