Floss: Wait for the Manager available before calling object manager
With some extreme cases, if chrome starts earlier than Floss manager,
chrome won't be able to register manager clients and the Bluetooth icon
will be gone in the system tray.

It is also observed that betty-chrome, login is blocked by the Floss
initialization callback. Therefore, we add a timeout to ensure that the
callback will be called as a workaround.

bluetooth is shown in the system tray.

Bug: 314699012
Test: stop ui; stop btmanagerd; start ui; start btmanagerd. Verify
Change-Id: Ie8f8601b29b1e179c73a7998f52e4c16f5a89504
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5758660
Commit-Queue: Yun-hao Chung <howardchung@chromium.org>
Reviewed-by: Ying Hsu <yinghsu@chromium.org>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Auto-Submit: Yun-hao Chung <howardchung@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337762}