Reset PIN attempt counter after changing it.
The original idea of the PIN attempt limit was that the client could
reregister to reset it. Reregistering requires interacting with Folsom,
which has hardware rate limiting. But we added the ability to change the
PIN in Chrome without reconsidering that plan.

Thus this change resets the PIN attempt counter whenever the user wraps
a new PIN. That requires GAIA reauthentication which meets that bar that
we have for changing the PIN, and thus should also be sufficient for
getting five more guess attempts at the enclave.

We have to consider whether this gives Google any extra powers too:

We can assume that Google can create valid RAPTs at will. Thus it could
give a device everything that it needs to make more guesses, but it
could already give a device the power to change the PIN and, from there,
the security domain secret can be extracted. We can assume that the
user's devices aren't cooperating with Google to undermine the user's
security.

Google could also create its own device in the user's account and get as
many PIN guesses as it wished with this change. But guessing the PIN
requires having a wrapped security domain secret, which we can assume
Google doesn't have because that would require Google having the
security domain secret in order to wrap it, which contradicts the
assumption that Google doesn't have that.

Tests for this change are in Chromium because that infrastructure makes
it easier to write a test for this case.

Bug: b/357603719
Change-Id: If7807f2c00e426c7ffb959e192e6a0bc7a2fb2c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5763169
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Auto-Submit: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337826}