[ios] Implement the delete action for tab groups panel
This CL implements the delete action in the context menu from a tab
groups panel cell.
If the group exists locally, it deletes a group and tabs in the group
locally, and deletes the group from the tab group sync service.
If the group doesn't exist locally, it deletes the group from the tab
group sync service.

Bug: 329629692
Change-Id: I61ca2f8320fbc744417088cb61be745947404ed5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5741503
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1337880}