FaceGaze: Turn off feature when DLC fails
FaceGaze will not work if the associated DLC fails, so this change
implements behavior to turn off FaceGaze when the DLC fails. This was
requested by UX as a follow-up to the DLC notifications behavior.

One thing to note is that this will make FaceGaze unusable in an
emulated environment because the DLC will either fail or the
contents of the DLC will fail to be read. However, FaceGaze never
truly worked in an emulated environment anyway so this should have no
impact for developers.

Bug: 309121742
Change-Id: Ibf57f36ae1b46bb0d3e5f29cb71c543fc9f896f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5757356
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Amanda Lin Dietz <aldietz@google.com>
Cr-Commit-Position: refs/heads/main@{#1337899}