Implement client-side contribution merging for Private Aggregation
This feature merges together histogram contributions with the same
bucket and filtering ID and drops contributions with a value of zero.
The contribution truncation step is also modified to account for this,
allowing for more utility out of the limit (i.e. fewer contributions
being dropped due to the limit).

This feature is not enabled by default, so this cl should be a no-op
until it is enabled.

Bug: 349980058
Change-Id: I357942f9901ab5f918a5b5473a8deffc87fe418e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5668046
Reviewed-by: Dan McArdle <dmcardle@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337916}