Check sync is inactive for disabled types when data is cleared by policy
Adding this CHECK to prevent data loss from sync server if
ClearBrowsingDataOnExitList or BrowsingDataLifetime policies run when
the sync server failed to deactivate  for the disabled data types.

For details on why a CHECK was chosen over an if condition, please see
comment#5 on the bug.

Bug: 41495797
Change-Id: Id172638ee7a7220be207b7319f23b832cac4deee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5757163
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Cr-Commit-Position: refs/heads/main@{#1337924}