Fix CHECK in DelegatedInkPointRendererGpu::StoreDelegatedInkPoint
The mentioned check was commented out because it was being hit and a
root cause had to be found.

When the pointer leaves the viewport a kMouseExited event is created.
The first pointer move event right after that exit event sometimes has
more than one calculated target down the pipeline in
EventProcessor::OnEventFromSource. From here, this event makes it down
to RenderWidgetHostImpl::ForwardDelegatedInkPoint, where we receive
two duplicate points. These points are forwarded to the GPU DelInk
renderer, and the second one causes a crash by hitting the CHECK
because the second point's timestamp is equal to the existing point,
instead of being larger.

This CL avoids hitting the CHECK and the unnecessary IPC call by
caching the last point sent and not making the call if the new point
is equal to the previous one.

Bug: 40784171
Change-Id: I214a7f831b2c0207a7f7d93642ba548b339b78bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5742013
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Sahir Vellani <sahir.vellani@microsoft.com>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1337963}