[iOS] Update TabGridMode, moving it to shared holder
This CL updates how the TabGridMode is handled in the Tab Grid. Instead
of having it passed through various means, this CL is creating a shared
holder for its value, accepting observers.

It means that any component that want to observer the current state of
the TabGridMode can use and observe this component. It allows to break
the command to switch modes and the complex update flows.

Low-Coverage-Reason: LARGE_SCALE_REFACTOR
Fixed: 357023309, 356590405
Change-Id: I7a3a84de1335ee8accab52a7154a5464c03dd0b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5756848
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Louis Romero <lpromero@google.com>
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337969}