Updating the `IpProtectionConfigCache` to utilize its current geo.
Before these changes, the config cache would check for and request
tokens based on the token manager's current geo. There is a chance where
the proxy list in the config cache has not refreshed to a newer geo. In
this case, we would like to check to see if the token cache manager may
have some leftover tokens from a previous token fill. By utilizing the
config cache's current geo (which is based on the proxy list), we are
able to optimize using all available tokens in this transitional phase.

Bug: 338975162
Change-Id: Ia632ddac8b259137c8f618d9cc72a7927fa4ca50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5763079
Reviewed-by: Andrew Williams <awillia@chromium.org>
Commit-Queue: Abhi Patel <abhipatel@chromium.org>
Reviewed-by: Dustin Mitchell <djmitche@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337983}