[A11y] When repairing prohibited name, avoid redundant description
Do not expose the prohibited name as a description if it would be
redundant with the inner contents of the object.

Also, implement ATK/IA2 name-from object attribute so that ATs can
make their own determination as to the value of the repaired name.
For more info about name-from, see:
https://docs.google.com/document/d/1-zjGmiOtjrHeBjXVGkhCcK2ya4Xq_k3yerHqSi_3qsY/edit#heading=h.benybc8eru3u

Bug: 350528330
Change-Id: Ibe2d4ccb777c24f0b5a471b8b45848051add0004
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3900708
Reviewed-by: Tim <tjudkins@chromium.org>
Reviewed-by: Brendon Tiszka <tiszka@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1338011}