Make cr-toggle focus visible in hcm
forced-color-adjust:none seemed to make a usable toggle in hcm with
dev tools rendering of forced-color active. In practice on Windows,
focus state was not visible, and the checked state was difficult to see.

Instead, I added forcing the knob color to be visible.

I left the comment about lack of specs, since I wasn't going off of a
spec here, and was just trying to make this usable.

Screencasts:
Before change:
https://screencast.googleplex.com/cast/NjYwOTI3MDQ3MTY1NTQyNHxlNDI1NDA3OS1lMg
After change:
https://screencast.googleplex.com/cast/NjM3NDMxODIxMTU5NjI4OHw1MzlkMzVmYS01OQ

Bug: 347774203
Change-Id: I490930448e65096645f4e8f9f2968a6c52c37404
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5756128
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Riley Tatum <rtatum@google.com>
Cr-Commit-Position: refs/heads/main@{#1338048}